move files into appropriate directories
