+++ One-Up/app/src/main/java/com/l3/one_up/adapters/FriendsAdapter.java
@@ -3,7 +3,6 @@ package com.l3.one_up.adapters;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,7 +11,7 @@ import android.widget.TextView;
 
 import com.l3.one_up.R;
 import com.l3.one_up.model.FacebookQuery;
-import com.l3.one_up.model.GlideApp;
+import com.l3.one_up.services.GlideApp;
 
 import java.util.ArrayList;
