+++ One-Up/app/src/main/java/com/l3/one_up/fragments/ActivitySelectionFragment.java
@@ -22,7 +22,7 @@ import android.view.ViewGroup;
 import android.widget.AutoCompleteTextView;
 import android.widget.TextView;
 
-import com.l3.one_up.Objective;
+import com.l3.one_up.model.Objective;
 import com.l3.one_up.R;
 
 import com.l3.one_up.adapters.ActivityItemAdapter;
