+++ One-Up/app/src/main/java/com/l3/one_up/fragments/FriendsFragment.java
@@ -31,8 +31,8 @@ import com.l3.one_up.adapters.FriendsAdapter;
 import com.l3.one_up.interfaces.FacebookCallComplete;
 import com.l3.one_up.model.Event;
 import com.l3.one_up.model.FacebookQuery;
-import com.l3.one_up.model.OrderFacebookUsersById;
-import com.l3.one_up.model.OrderParseUsersByFbId;
+import com.l3.one_up.services.OrderFacebookUsersById;
+import com.l3.one_up.services.OrderParseUsersByFbId;
 import com.l3.one_up.model.PowerUp;
 import com.l3.one_up.model.User;
 import com.parse.FindCallback;
