+++ One-Up/app/src/main/java/com/l3/one_up/fragments/ProfileFragment.java
@@ -1,16 +1,13 @@
 package com.l3.one_up.fragments;
 
 import android.app.Activity;
-import android.app.FragmentTransaction;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -18,7 +15,7 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.l3.one_up.DeepLinkingActivity;
+import com.l3.one_up.services.DeepLinkingActivity;
 import com.l3.one_up.R;
 import com.parse.ParseUser;
