--- One-Up/app/src/main/java/com/l3/one_up/model/GlideAppModule.java
@@ -1,12 +1,0 @@
-package com.l3.one_up.model;
-
-import com.bumptech.glide.annotation.GlideModule;
-import com.bumptech.glide.module.AppGlideModule;
-
-/**
- * Created by luzcamacho on 7/25/18.
- */
-
-@GlideModule
-public class GlideAppModule extends AppGlideModule {
-}
