+++ One-Up/app/src/main/java/com/l3/one_up/model/Objective.java
@@ -1,4 +1,4 @@
-package com.l3.one_up;
+package com.l3.one_up.model;
 
 public enum Objective {
     GOAL,
