+++ One-Up/app/src/main/java/com/l3/one_up/services/MyAppGlideModule.java
@@ -1,5 +1,7 @@
-package com.l3.one_up;
+package com.l3.one_up.services;
 
+import com.bumptech.glide.annotation.GlideModule;
 import com.bumptech.glide.module.AppGlideModule;
 
+@GlideModule
 public class MyAppGlideModule extends AppGlideModule { }
