+++ One-Up/app/src/main/res/layout/activity_deep_linking.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.l3.one_up.DeepLinkingActivity">
+    tools:context="com.l3.one_up.services.DeepLinkingActivity">
 
     <com.facebook.login.widget.LoginButton
         android:id="@+id/btFBConnect"
